Update the README
The README had gotten a bit out of date in several places. Probably the most noticeable was the discussion of on-line tools. ShareLaTeX doesn't exist anymore (they merged with Overleaf), and almost everyone uses Overleaf, so it seemed to make sense to put it first.

We also still had references to `sig-alternate.cls`, which we haven't used in years, so I deleted those.